+++ changelog.md
@@ -14,6 +14,7 @@
 - [#3105](https://github.com/ignite/cli/pull/3105) Addition of `ignite plugin describe <path>` command
 - [#2995](https://github.com/ignite/cli/pull/2995/) Add `ignite network request remove-validator` command.
 - [#2999](https://github.com/ignite/cli/pull/2999/) Add `ignite network request remove-account` command.
+- [#2458](https://github.com/ignite/cli/issues/2458) New `chain serve` command UI.
 
 ### Changes